core/requirements.txt
playwright>=1.45
pillow>=10.0.0
reportlab>=4.0.0
pypdf>=4.2.0
tkinterdnd2>=0.4.2
customtkinter>=5.2.0
